Add function to clear a bit at an index
Clear_bit uses the AND and NOT operations to clear the bit at an index.
A bit mask equal to 1L is left shifted by 'index' bits to the index posiition.
The NOT operation negates the bits in all positions then performs an AND operation to negate thhe bit at the index.